requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.1
imgaug==0.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
PyTurboJPEG==1.7.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sympy==1.12.1
tensorboardX==2.6.2.2
termcolor==2.4.0
tifffile==2024.5.22
timm==1.0.3
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.28
ultralytics-thop==0.2.7
urllib3==2.2.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
